fix: local fake data response
