+++ src/app.module.ts
@@ -11,6 +11,7 @@ import { DatabaseModule } from './domain/database/database.module';
 import { QuestionBankModule } from './domain/questionsBank/questionBank.module';
 import { QuestionBankService } from './domain/questionsBank/service/questionBank.service';
 import { QuestionBankRepository } from './domain/questionsBank/repository/questionsBank.repository';
+import { BatchService } from './domain/batch/batch.service';
 
 @Module({
   imports: [
@@ -27,6 +28,11 @@ import { QuestionBankRepository } from './domain/questionsBank/repository/questi
     QuestionBankModule,
   ],
   controllers: [PromptController],
-  providers: [PromptService, QuestionBankService, QuestionBankRepository],
+  providers: [
+    PromptService,
+    QuestionBankService,
+    QuestionBankRepository,
+    BatchService,
+  ],
 })
 export class AppModule {}
